implemented betting system

+++ src/main/scala/de/htwg/se/blackjackKN/model/Bet.scala
@@ -1,4 +1,3 @@
 package de.htwg.se.blackjackKN.model
 
-case class Bet(value: Int) {
-}
+case class Bet(value: Double) {}
